# backend/public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "ToDo",
        "contact": {
            "name": "Ahmed Helal Ahmed",
            "url": "https://www.facebook.com/A.Helal.A.A"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Get All tasks"
                ],
                "summary": "Get tasks",
                "description": "Display a list of the tasks",
                "responses": {
                    "200": {
                        "description": "Successful operation - All data received"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Create a new task"
                ],
                "summary": "Store a task.",
                "description": "Store the task in database and return it in JSON",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Task description",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Task to do"
                        },
                        "type": "string"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "description": "Task details",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "This task is important because . . ."
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation - task stored"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    }
                }
            }
        },
        "/tasks/{ID}": {
            "get": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Get a task",
                    "Show a task"
                ],
                "summary": "show a task by ID.",
                "description": "Get the task that has ID in the URL in JSON",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "task's id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    },
                    "404": {
                        "description": "Task Model NOT Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Update a tasks"
                ],
                "summary": "Update a task by ID.",
                "description": "Update the task that has the ID in the URL",
                "parameters": [
                    {
                        "name": "description",
                        "in": "body",
                        "description": "Description of the task",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Task to do"
                        },
                        "type": "string"
                    },
                    {
                        "name": "details",
                        "in": "body",
                        "description": "Details of the task",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "This task is important because . . ."
                        },
                        "type": "string"
                    },
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "ID of a task",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation - Task updated"
                    },
                    "404": {
                        "description": "Task Model NOT Found"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Delete a task"
                ],
                "summary": "Delete a task.",
                "description": "Delete the task that has the ID in the URL",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "task's id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "204": {
                        "description": "Successful operation - NO CONTENT"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    }
                }
            }
        },
        "/tasks/{ID}/edit": {
            "get": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Edit a task",
                    "Get a task"
                ],
                "summary": "Edit a task by ID.",
                "description": "Edit - return the task that has the ID in the URL in JSON",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "task's id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    },
                    "404": {
                        "description": "Task Model NOT Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/tasks/{ID1}/{ID2}": {
            "post": {
                "tags": [
                    "Tasks - All CRUD operations -",
                    "Swap tasks data"
                ],
                "summary": "Swap the IDs of two tasks.",
                "description": "Swap two tasks data (description and details) which have the ids ID1 and ID2 in the URL and return the new updated data in JSON",
                "parameters": [
                    {
                        "name": "ID1",
                        "in": "path",
                        "description": "first task's id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "ID2",
                        "in": "path",
                        "description": "second task's id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "400": {
                        "description": "Bad operation - Fail to pass validation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "403": {
                        "description": "Forbidden - Authenticated, but does not have the permissions to perform an action"
                    }
                }
            }
        }
    },
    "definitions": {
        "Task": {
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "description": "the name of a task",
                    "type": "string"
                },
                "details": {
                    "description": "the details of a task",
                    "type": "string"
                }
            },
            "default": false
        }
    }
}